+++ java/src/main/java/io/cucumber/cucumberexpressions/CucumberExpressionParser.java
@@ -1,28 +1,27 @@
 package io.cucumber.cucumberexpressions;
 
-import io.cucumber.cucumberexpressions.Ast.Node;
 import io.cucumber.cucumberexpressions.Ast.Token;
-import io.cucumber.cucumberexpressions.Ast.Token.TokenType;
+import io.cucumber.cucumberexpressions.Ast.Token.Type;
 import org.apiguardian.api.API;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import static io.cucumber.cucumberexpressions.Ast.Node.NodeType.ALTERNATION_NODE;
-import static io.cucumber.cucumberexpressions.Ast.Node.NodeType.ALTERNATIVE_NODE;
-import static io.cucumber.cucumberexpressions.Ast.Node.NodeType.EXPRESSION_NODE;
-import static io.cucumber.cucumberexpressions.Ast.Node.NodeType.OPTIONAL_NODE;
-import static io.cucumber.cucumberexpressions.Ast.Node.NodeType.PARAMETER_NODE;
-import static io.cucumber.cucumberexpressions.Ast.Node.NodeType.TEXT_NODE;
-import static io.cucumber.cucumberexpressions.Ast.Token.TokenType.ALTERNATION;
-import static io.cucumber.cucumberexpressions.Ast.Token.TokenType.BEGIN_OPTIONAL;
-import static io.cucumber.cucumberexpressions.Ast.Token.TokenType.BEGIN_PARAMETER;
-import static io.cucumber.cucumberexpressions.Ast.Token.TokenType.END_OF_LINE;
-import static io.cucumber.cucumberexpressions.Ast.Token.TokenType.END_OPTIONAL;
-import static io.cucumber.cucumberexpressions.Ast.Token.TokenType.END_PARAMETER;
-import static io.cucumber.cucumberexpressions.Ast.Token.TokenType.START_OF_LINE;
-import static io.cucumber.cucumberexpressions.Ast.Token.TokenType.WHITE_SPACE;
+import static io.cucumber.cucumberexpressions.Node.Type.ALTERNATION_NODE;
+import static io.cucumber.cucumberexpressions.Node.Type.ALTERNATIVE_NODE;
+import static io.cucumber.cucumberexpressions.Node.Type.EXPRESSION_NODE;
+import static io.cucumber.cucumberexpressions.Node.Type.OPTIONAL_NODE;
+import static io.cucumber.cucumberexpressions.Node.Type.PARAMETER_NODE;
+import static io.cucumber.cucumberexpressions.Node.Type.TEXT_NODE;
+import static io.cucumber.cucumberexpressions.Ast.Token.Type.ALTERNATION;
+import static io.cucumber.cucumberexpressions.Ast.Token.Type.BEGIN_OPTIONAL;
+import static io.cucumber.cucumberexpressions.Ast.Token.Type.BEGIN_PARAMETER;
+import static io.cucumber.cucumberexpressions.Ast.Token.Type.END_OF_LINE;
+import static io.cucumber.cucumberexpressions.Ast.Token.Type.END_OPTIONAL;
+import static io.cucumber.cucumberexpressions.Ast.Token.Type.END_PARAMETER;
+import static io.cucumber.cucumberexpressions.Ast.Token.Type.START_OF_LINE;
+import static io.cucumber.cucumberexpressions.Ast.Token.Type.WHITE_SPACE;
 import static io.cucumber.cucumberexpressions.CucumberExpressionException.createAlternationNotAllowedInOptional;
 import static io.cucumber.cucumberexpressions.CucumberExpressionException.createInvalidParameterTypeName;
 import static io.cucumber.cucumberexpressions.CucumberExpressionException.createMissingEndToken;
@@ -200,9 +199,9 @@ public final class CucumberExpressionParser {
     }
 
     private static Parser parseBetween(
-            Node.NodeType type,
-            TokenType beginToken,
-            TokenType endToken,
+            Node.Type type,
+            Type beginToken,
+            Type endToken,
             List<Parser> parsers) {
         return (expression, tokens, current) -> {
             if (!lookingAt(tokens, current, beginToken)) {
@@ -228,7 +227,7 @@ public final class CucumberExpressionParser {
             List<Parser> parsers,
             List<Token> tokens,
             int startAt,
-            TokenType... endTokens) {
+            Type... endTokens) {
         int current = startAt;
         int size = tokens.size();
         List<Node> ast = new ArrayList<>();
@@ -262,8 +261,8 @@ public final class CucumberExpressionParser {
         throw new IllegalStateException("No eligible parsers for " + tokens);
     }
 
-    private static boolean lookingAtAny(List<Token> tokens, int at, TokenType... tokenTypes) {
-        for (TokenType tokeType : tokenTypes) {
+    private static boolean lookingAtAny(List<Token> tokens, int at, Type... tokenTypes) {
+        for (Type tokeType : tokenTypes) {
             if (lookingAt(tokens, at, tokeType)) {
                 return true;
             }
@@ -271,7 +270,7 @@ public final class CucumberExpressionParser {
         return false;
     }
 
-    private static boolean lookingAt(List<Token> tokens, int at, TokenType token) {
+    private static boolean lookingAt(List<Token> tokens, int at, Type token) {
         if (at < 0) {
             // If configured correctly this will never happen
             // Keep for completeness
