+++ java/src/test/java/io/cucumber/cucumberexpressions/CucumberExpressionTokenizerTest.java
@@ -1,7 +1,7 @@
 package io.cucumber.cucumberexpressions;
 
 import io.cucumber.cucumberexpressions.Ast.Token;
-import io.cucumber.cucumberexpressions.Ast.Token.TokenType;
+import io.cucumber.cucumberexpressions.Ast.Token.Type;
 import org.junit.jupiter.api.extension.ParameterContext;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.converter.ArgumentConversionException;
@@ -77,7 +77,7 @@ class CucumberExpressionTokenizerTest {
 
     static class YamlableToken {
         public String text;
-        public TokenType type;
+        public Type type;
         public int start;
         public int end;
